um: dbm: Add apphint for DRM driver name
Enable support for an Apphint "DbmDriverName=<string>" using which
an user can indicate an application which DRM driver to use to
allocate GEM buffers.

This feature is useful when there are multiple DRM drivers
in the system, e.g., omapdrm and vdrm coexistence

If AppHint is not provided, the default driver name
DBM_DRIVER_NAME_STRING (= omapdrm) indicated during build is used.

Signed-off-by: Subhajit Paul <subhajit_paul@ti.com>
Signed-off-by: Ramprasad N <x0038811@ti.com>