all: rebuild with major fixes applied
Rebuild with fixes for several conformance issues, allocation bugs, and
other oddities. This addresses several issues with things like glmark2
vec4 operations, NPOT textures, and devices with a PRIMARY and RENDER
display node.

This currently has a GLES2 conformance pass rate of 94.2% on AM57X,
which may sound a little low for a release, but this is a substantial
improvement over the previously unknown pass rate.

Signed-off-by: Randolph Sapp <rs@ti.com>